backend/requirements/base.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
django-cors-headers
anthropic
gunicorn
whitenoise
dj-database-url

